Add Address model, repository, and service; update user registration to include address
Combined Migration V1 through V7 as the new V1 file

+++ src/main/java/com/ICS499/Application/model/Order.java
@@ -15,7 +15,7 @@ public class Order {
         this.payment = pay;
      }
 
-     public void addCustomer(String name, String streetAddress,String city, String state, int zipCode, String phoneNum, String email ){
+     public void addCustomer(String name, String streetAddress,String city, String state, Long zipCode, String phoneNum, String email ){
         Customer customer = new Customer(name, streetAddress, city, state, zipCode, phoneNum, email);
      }
